Add minTrade parameter
This activates an alternate fitness algoritim until the minimum wins are met.  after they are met it switches to the normal fitness algoritim to find the best win/loss and profit.

This can help prevent situations where the best tactic FOUND is to don't trade.  Or you just want to have more influence on its evolution.